Moved drive into main directory.
