--- pygsheets/api/__init__.py
@@ -1,1 +1,0 @@
-from pygsheets.api.drive import DriveAPIWrapper
